requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.7  # For PostgreSQL
Pillow==10.0.1
dj-database-url==2.2.0  # For Render's PostgreSQL
whitenoise==6.7.0  # For serving static files
gunicorn==22.0.0  # For Render's WSGI server